Add fuzzer for series parser
