+++ fuzzers/CMakeLists.txt
@@ -1,5 +1,7 @@
 include_directories(../libakumuli)
 
+# Compression
+
 add_executable(afl_compression
     afl_compression.cpp
     ../libakumuli/storage_engine/compression.cpp
@@ -13,3 +15,23 @@ target_link_libraries(afl_compression
 )
 
 set_target_properties(afl_compression PROPERTIES EXCLUDE_FROM_ALL 1)
+
+# Series name parser
+
+add_executable(afl_series_name_parser
+    afl_series_name_parser.cpp
+    ../libakumuli/index/seriesparser.cpp
+    ../libakumuli/index/stringpool.cpp
+    ../libakumuli/index/invertedindex.cpp
+    ../libakumuli/util.cpp
+    ../libakumuli/status_util.cpp
+    ../libakumuli/log_iface.cpp
+    )
+
+target_link_libraries(afl_series_name_parser
+    pthread
+    "${Boost_LIBRARIES}"
+    "${APR_LIBRARY}"
+)
+
+set_target_properties(afl_series_name_parser PROPERTIES EXCLUDE_FROM_ALL 1)
